+++ source/gameos/include/mfcplatform.hpp
@@ -17,6 +17,6 @@
 NTSTATUS	__stdcall ProcessException( PEXCEPTION_POINTERS pep);
 void		__stdcall InitExceptionHandler( PSTR pCmdLine);
 void		__stdcall InitGameOS( HINSTANCE hInstance, HWND hWindow, PSTR pCmdLine);
-LRESULT		__stdcall GameOSWinProc( HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam );
-ULONG		__stdcall RunGameOSLogic(void);
+LRESULT		__stdcall GameOSWinProc( HWND hWnd, uint32_t uMsg, WPARAM wParam, LPARAM lParam );
+uint32_t		__stdcall RunGameOSLogic(void);
 void		__stdcall ExitGameOS(void);
